All my old commits from the previous project repository (which became corrupted unfortunately)
These are all the commits (in png form) that were in my previous repository that used to hold this project.  Unfortunately after I changed the folder name and file name (shown in my last commit message) the entire project stopped working so I had to move the .java files into a new project and new repository as a result.  After doing all that stuff, I found out the project STILL wasnt working due to a weird exception being thrown.  A short while ago, I found out it was because I never included the icon image for the bank app and thats what broke the original project.  Anyway, it works now and I've included every important file in this repository!